Improve Formatter
This commit changes the approach that we use to format the code.

Now, instead of relying on "{:text, "\n", _meta}" generated by
HTMLTokenizer, we are discarting it and checking if a line break
is supposed to be added according to the previous tag.

This makes the module `EnsureLineBreaks` obsolete since we don't need
use line breaks anymore.