Refactor Formatter (#49)
This PRs refactor the formatter code to use [Inspect.Alegbra](https://hexdocs.pm/elixir/Inspect.Algebra.html). 

Besides improving the code, it make things a lot easier since we don't 
need to manipulate indention manually as we were doing before.

Also, it improves how we have been formatting some specific situations. 
For instance, before the formatter would always break a line between 
block tags with attributes:

```html
<p class="foo">Text</p>

# would be formatted to:
<p class="foo">
  Text
</p>
```

Now with `Algebra`, it will only break to lines in case it doesn't fit 
in the current line. In the case mentioned above it will keep as is.

This refactor is also supposed to fix some issues we had with comments 
and script tags. 

# TODO
- [x] Build a HTML tree
  - [x] basic HTML structure
  - [x] self closed tags
  - [x] eex expressions 
  - [x] case/cond (-->)
- [x] Format the HTML tree using Algebra module
  - [x] basic HTML
  - [x] HTML attributes
  - [x] Format Eex blocks
  - [x] Handle inline self close tag 
  - [x] eex expressions. We should use Elixir format for expressions that starts and finishes at the same line.
  - [x] Handle long texts along with inline elements
  - [x] Format multiple expressions in HTML attributes
  - [x] Keep intentional line breaks
  - [x] Handle tags we don't want to format such as script, styles, comments
  - [x] Handle pre,code, etc... tags that we don't want to format at all